refactor: compress dirs
