+++ packages/x6/src/v1/core/port.ts
@@ -2,8 +2,8 @@ import { Size, KeyValue } from '../../types'
 import { Point, Rectangle } from '../../geometry'
 import { JSONObject, ObjectExt } from '../../util'
 import { Attr } from '../attr'
-import { PortLayout } from '../port-layout'
-import { PortLabelLayout } from '../port-label-layout'
+import { PortLayout } from '../port/port-layout'
+import { PortLabelLayout } from '../port/port-label-layout'
 import { PortLayoutRegistry, PortLabelLayoutRegistry } from '../registry'
 import { Markup } from './markup'
