+++ packages/x6/src/v1/port/port-label-layout/inout.ts
@@ -1,4 +1,4 @@
-import { Point, Rectangle } from '../../geometry'
+import { Point, Rectangle } from '../../../geometry'
 import { PortLabelLayout } from './index'
 import { toResult } from './util'
