+++ packages/x6/src/v1/port/port-label-layout/radial.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../geometry'
+import { Point } from '../../../geometry'
 import { PortLabelLayout } from './index'
 import { toResult } from './util'
