+++ packages/x6/src/v1/port/port-layout/util.ts
@@ -1,5 +1,5 @@
-import { NumberExt } from '../../util'
-import { Point, Rectangle } from '../../geometry'
+import { NumberExt } from '../../../util'
+import { Point, Rectangle } from '../../../geometry'
 import { PortLayout } from './index'
 
 export function normalizePoint(
