+++ packages/x6/src/v1/registry/grid.ts
@@ -1,4 +1,4 @@
-import { Grid } from '../grid'
+import { Grid } from '../style'
 import { Registry } from './registry'
 
 export const GridRegistry = Registry.create<
