+++ packages/x6/src/v1/registry/index.ts
@@ -1,3 +1,5 @@
+export { Registry } from './registry'
+
 export * from './node'
 export * from './edge'
 export * from './view'
