+++ packages/x6/src/v1/style/background/flip-x.ts
