+++ packages/x6/src/v1/style/background/flip-y.ts
