+++ packages/x6/src/v1/style/background/index.ts
@@ -1,7 +1,7 @@
 import { ValuesType } from 'utility-types'
 import * as CSS from 'csstype'
 import * as patterns from './index-rollup'
-import { KeyValue } from '../../types'
+import { KeyValue } from '../../../types'
 
 export namespace Background {
   export interface Options {
