+++ packages/x6/src/v1/style/filter/index.ts
@@ -1,5 +1,5 @@
 import { NonUndefined } from 'utility-types'
-import { KeyValue } from '../../types'
+import { KeyValue } from '../../../types'
 import * as filters from './index-rollup'
 
 export namespace Filter {
