+++ packages/x6/src/v1/style/filter/invert.ts
