+++ packages/x6/src/v1/style/filter/util.ts
