+++ packages/x6/src/v1/style/highlighter/index.ts
@@ -1,5 +1,5 @@
-import { KeyValue } from '../../types'
-import { CellView } from '../core'
+import { KeyValue } from '../../../types'
+import { CellView } from '../../core'
 import * as highlighters from './index-rollup'
 
 export namespace Highlighter {
